Revert "CspaceFreeBox find separation certificates for a given box. (#19742)"
This reverts commit 7904810e90b91fdef2bef54d7a5d806940165626.
