+++ geometry/optimization/cspace_free_box.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <optional>
-#include <unordered_map>
 #include <vector>
 
 #include "drake/geometry/optimization/cspace_free_polytope_base.h"
@@ -156,31 +154,6 @@ class CspaceFreeBox : public CspaceFreePolytopeBase {
                 SeparatingPlaneOrder plane_order,
                 const Options& options = Options{});
 
-  /** Finds the certificates that the C-space box {q | q_box_lower <= q <=
-   * q_box_upper} is collision free.
-   *
-   * @param q_box_lower The lower bound of the C-space box.
-   * @param q_box_upper The upper bound of the C-space box.
-   * @param ignored_collision_pairs We ignore the pair of geometries in
-   * `ignored_collision_pairs`.
-   * @param[out] certificates Contains the certificate we successfully found for
-   * each pair of geometries. Notice that depending on `options`, the program
-   * could search for the certificate for each geometry pair in parallel, and
-   * will terminate the search once it fails to find the certificate for any
-   * pair. At termination, the pair of geometries whose optimization hasn't been
-   * finished will not show up in @p certificates.
-   * @retval success If true, then we have certified that the C-space box
-   * {q | q_box_lower<=q<=q_box_upper} is collision free. Otherwise
-   * success=false.
-   */
-  bool FindSeparationCertificateGivenBox(
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_lower,
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_upper,
-      const IgnoredCollisionPairs& ignored_collision_pairs,
-      const FindSeparationCertificateOptions& options,
-      std::unordered_map<SortedPair<geometry::GeometryId>,
-                         SeparationCertificateResult>* certificates) const;
-
  private:
   // Forward declare the tester class that will test the private members.
   friend class CspaceFreeBoxTester;
@@ -238,37 +211,6 @@ class CspaceFreeBox : public CspaceFreePolytopeBase {
       const PlaneSeparatesGeometries& plane_geometries,
       const VectorX<symbolic::Polynomial>& s_minus_s_lower,
       const VectorX<symbolic::Polynomial>& s_upper_minus_s) const;
-
-  /*
-   Finds the certificates that the C-space box {q | q_box_lower <= q <=
-   q_box_upper} is collision free.
-   @retval certificates certificates[i] is the separation certificate for a pair
-   of geometries. If we cannot certify or haven't certified the separation for
-   this pair, then certificates[i] contains std::nullopt. Note that when we run
-   this function in parallel and options.terminate_at_failure=true, we will
-   terminate all the remaining certification programs that have been launched,
-   so certificates[i] = std::nullopt could be either because that we have
-   attempted to find the certificate for this pair of geometry but failed, or it
-   could be that we fail to find the certificate for another pair and haven't
-   attempted to find the certificate for this pair.
-   The geometry pair which certificates[i] certifies is given by
-   separating_planes()[certificates[i].plane_index].geometry_pair().
-   */
-  [[nodiscard]] std::vector<std::optional<SeparationCertificateResult>>
-  FindSeparationCertificateGivenBox(
-      const IgnoredCollisionPairs& ignored_collision_pairs,
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_lower,
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_upper,
-      const FindSeparationCertificateOptions& options) const;
-
-  /*
-   Adds the constraint that each column of s_inner_pts is in the box s_box_lower
-   <= s <= s_box_upper.
-   */
-  void AddCspaceBoxContainment(solvers::MathematicalProgram* prog,
-                               const VectorX<symbolic::Variable>& s_box_lower,
-                               const VectorX<symbolic::Variable>& s_box_upper,
-                               const Eigen::MatrixXd& s_inner_pts) const;
 };
 }  // namespace optimization
 }  // namespace geometry
