+++ geometry/optimization/test/c_iris_test_utilities.h
@@ -240,17 +240,6 @@ class CspaceFreeBoxTester {
         plane_geometries, s_minus_s_lower, s_upper_minus_s);
   }
 
-  [[nodiscard]] std::vector<
-      std::optional<CspaceFreeBox::SeparationCertificateResult>>
-  FindSeparationCertificateGivenBox(
-      const CspaceFreeBox::IgnoredCollisionPairs& ignored_collision_pairs,
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_lower,
-      const Eigen::Ref<const Eigen::VectorXd>& q_box_upper,
-      const FindSeparationCertificateOptions& options) const {
-    return cspace_free_box_.FindSeparationCertificateGivenBox(
-        ignored_collision_pairs, q_box_lower, q_box_upper, options);
-  }
-
  private:
   CspaceFreeBox cspace_free_box_;
 };
