+++ geometry/optimization/test/cspace_free_box_with_mosek_test.cc
@@ -1,60 +1,14 @@
-#include <limits>
-#include <queue>
-
 #include "drake/common/fmt_eigen.h"
 #include "drake/common/test_utilities/eigen_matrix_compare.h"
 #include "drake/common/test_utilities/symbolic_test_util.h"
 #include "drake/geometry/optimization/cspace_free_box.h"
 #include "drake/geometry/optimization/test/c_iris_test_utilities.h"
-#include "drake/multibody/inverse_kinematics/inverse_kinematics.h"
 #include "drake/solvers/mosek_solver.h"
-#include "drake/solvers/solve.h"
 
 namespace drake {
 namespace geometry {
 namespace optimization {
 namespace {
-const double kInf = std::numeric_limits<double>::infinity();
-// Only use two threads during testing. This value should match the "cpu" tag
-// in BUILD.bazel defining this test.
-constexpr int kTestConcurrency = 2;
-
-// Generate a matrix where each row is a grid point in the box.
-Eigen::MatrixXd CalcBoxGrid(const Eigen::Ref<const Eigen::VectorXd>& box_lower,
-                            const Eigen::Ref<const Eigen::VectorXd>& box_upper,
-                            const std::vector<int>& grid_size) {
-  const int x_dim = box_lower.rows();
-  DRAKE_DEMAND(box_upper.rows() == x_dim);
-  DRAKE_DEMAND(ssize(grid_size) == x_dim);
-  std::vector<Eigen::VectorXd> x_grid;
-  for (int i = 0; i < x_dim; ++i) {
-    x_grid.push_back(
-        Eigen::VectorXd::LinSpaced(grid_size[i], box_lower(i), box_upper(i)));
-  }
-  std::queue<std::vector<double>> grid_pts;
-  grid_pts.emplace();
-  for (int i = 0; i < x_dim; ++i) {
-    const int queue_size = grid_pts.size();
-    for (int j = 0; j < queue_size; ++j) {
-      for (int k = 0; k < grid_size[i]; ++k) {
-        std::vector<double> grid = grid_pts.front();
-        grid.push_back(x_grid[i](k));
-        grid_pts.push(grid);
-      }
-      grid_pts.pop();
-    }
-  }
-  Eigen::MatrixXd ret(grid_pts.size(), x_dim);
-  int pt_count = 0;
-  while (!grid_pts.empty()) {
-    ret.row(pt_count) =
-        Eigen::Map<Eigen::RowVectorXd>(grid_pts.front().data(), x_dim);
-    pt_count++;
-    grid_pts.pop();
-  }
-  return ret;
-}
-
 // @param q_samples Each row of s_samples is a sampled configuration q.
 // @param a Maps the plane index to the separating plane parameter `a` in {x|
 // aᵀx+b=0}
@@ -108,25 +62,6 @@ void CheckSosLagrangians(
       EXPECT_TRUE(IsPolynomialSos(lagrangians.s_box_lower()(i), 0.0));
       EXPECT_TRUE(IsPolynomialSos(lagrangians.s_box_upper()(i), 0.0));
     }
-  }
-}
-
-std::optional<Eigen::VectorXd> FindInCollisionPosture(
-    const multibody::MultibodyPlant<double>& plant,
-    const SortedPair<geometry::GeometryId>& geometry_pair,
-    systems::Context<double>* plant_context,
-    const Eigen::Ref<const Eigen::VectorXd>& q_box_lower,
-    const Eigen::Ref<const Eigen::VectorXd>& q_box_upper,
-    const std::optional<Eigen::VectorXd>& q_guess) {
-  multibody::InverseKinematics ik(plant, plant_context);
-  ik.get_mutable_prog()->AddBoundingBoxConstraint(q_box_lower, q_box_upper,
-                                                  ik.q());
-  ik.AddDistanceConstraint(geometry_pair, -kInf, 0);
-  const auto result = solvers::Solve(ik.prog(), q_guess, std::nullopt);
-  if (result.is_success()) {
-    return result.GetSolution(ik.q());
-  } else {
-    return std::nullopt;
   }
 }
 
@@ -293,138 +228,6 @@ TEST_F(CIrisToyRobotTest, ConstructPlaneSearchProgram) {
         q_samples.rows(), q_box_lower(1), q_box_upper(1));
 
     check_certificate(separation_certificate_program, result, q_samples);
-  }
-}
-
-TEST_F(CIrisToyRobotTest, FindSeparationCertificateGivenBoxSuccess) {
-  // Test FindSeparationCertificateGivenBox that successfully finds the
-  // certificates.
-  CspaceFreeBoxTester tester(plant_, scene_graph_,
-                             SeparatingPlaneOrder::kAffine);
-  const Eigen::VectorXd q_position_lower = plant_->GetPositionLowerLimits();
-  const Eigen::VectorXd q_position_upper = plant_->GetPositionUpperLimits();
-  const Eigen::VectorXd q_box_lower =
-      0.8 * q_position_lower + 0.2 * q_position_upper;
-  const Eigen::VectorXd q_box_upper =
-      0.7 * q_position_lower + 0.3 * q_position_upper;
-  FindSeparationCertificateOptions options;
-  const CspaceFreeBox::IgnoredCollisionPairs ignored_collision_pairs{
-      SortedPair<geometry::GeometryId>(world_box_, body2_sphere_)};
-  Eigen::VectorXd s_box_lower;
-  Eigen::VectorXd s_box_upper;
-  Eigen::VectorXd q_star;
-  tester.ComputeSBox(q_box_lower, q_box_upper, &s_box_lower, &s_box_upper,
-                     &q_star);
-  for (int num_threads : {1, kTestConcurrency}) {
-    options.num_threads = num_threads;
-    options.terminate_at_failure = false;
-    const auto certificates_result = tester.FindSeparationCertificateGivenBox(
-        ignored_collision_pairs, q_box_lower, q_box_upper, options);
-    EXPECT_EQ(certificates_result.size(),
-              tester.cspace_free_box().separating_planes().size() -
-                  ignored_collision_pairs.size());
-    // The C-space box is collision free.
-    EXPECT_TRUE(std::all_of(
-        certificates_result.begin(), certificates_result.end(),
-        [](const std::optional<CspaceFreeBox::SeparationCertificateResult>&
-               result) {
-          return result.has_value();
-        }));
-    // Make sure we really find the separation plane.
-    const Eigen::MatrixXd q_samples =
-        CalcBoxGrid(q_box_lower, q_box_upper, {10, 10, 10});
-    for (const auto& certificate_result : certificates_result) {
-      const auto& separating_plane =
-          tester.cspace_free_box()
-              .separating_planes()[certificate_result->plane_index];
-      CheckSeparationBySamples(
-          tester, *(this->diagram_), q_samples, certificate_result->a,
-          certificate_result->b, q_star,
-          SortedPair<geometry::GeometryId>(
-              separating_plane.positive_side_geometry->id(),
-              separating_plane.negative_side_geometry->id()));
-      CheckSosLagrangians(
-          certificate_result->positive_side_rational_lagrangians);
-      CheckSosLagrangians(
-          certificate_result->negative_side_rational_lagrangians);
-    }
-    // Now test the public FindSeparationCertificateGivenBox function.
-    std::unordered_map<SortedPair<geometry::GeometryId>,
-                       CspaceFreeBox::SeparationCertificateResult>
-        certificates;
-    const bool is_success =
-        tester.cspace_free_box().FindSeparationCertificateGivenBox(
-            q_box_lower, q_box_upper, ignored_collision_pairs, options,
-            &certificates);
-    EXPECT_TRUE(is_success);
-    EXPECT_EQ(certificates.size(),
-              tester.cspace_free_box().separating_planes().size() -
-                  ignored_collision_pairs.size());
-  }
-}
-
-TEST_F(CIrisToyRobotTest, FindSeparationCertificateGivenBoxFailure) {
-  // Test FindSeparationCertificateGivenBox but fails.
-  CspaceFreeBoxTester tester(plant_, scene_graph_,
-                             SeparatingPlaneOrder::kAffine);
-  const Eigen::VectorXd q_position_lower = plant_->GetPositionLowerLimits();
-  const Eigen::VectorXd q_position_upper = plant_->GetPositionUpperLimits();
-  const Eigen::VectorXd q_box_lower =
-      0.8 * q_position_lower + 0.2 * q_position_upper;
-  const Eigen::VectorXd q_box_upper =
-      0.1 * q_position_lower + 0.9 * q_position_upper;
-  FindSeparationCertificateOptions options;
-  const CspaceFreeBox::IgnoredCollisionPairs ignored_collision_pairs{
-      SortedPair<geometry::GeometryId>(world_box_, body2_sphere_)};
-  Eigen::VectorXd s_box_lower;
-  Eigen::VectorXd s_box_upper;
-  Eigen::VectorXd q_star;
-  tester.ComputeSBox(q_box_lower, q_box_upper, &s_box_lower, &s_box_upper,
-                     &q_star);
-  const Eigen::MatrixXd q_samples =
-      CalcBoxGrid(q_box_lower, q_box_upper, {10, 10, 10});
-  for (int num_threads : {1, kTestConcurrency}) {
-    options.num_threads = num_threads;
-    options.terminate_at_failure = false;
-    std::unordered_map<SortedPair<geometry::GeometryId>,
-                       CspaceFreeBox::SeparationCertificateResult>
-        certificates;
-    const bool success =
-        tester.cspace_free_box().FindSeparationCertificateGivenBox(
-            q_box_lower, q_box_upper, ignored_collision_pairs, options,
-            &certificates);
-    EXPECT_FALSE(success);
-    EXPECT_LT(certificates.size(),
-              tester.cspace_free_box().separating_planes().size() -
-                  ignored_collision_pairs.size());
-    for (const auto& separating_plane :
-         tester.cspace_free_box().separating_planes()) {
-      const SortedPair<geometry::GeometryId> geometry_pair{
-          separating_plane.positive_side_geometry->id(),
-          separating_plane.negative_side_geometry->id()};
-      if (ignored_collision_pairs.count(geometry_pair) == 0) {
-        auto it = certificates.find(geometry_pair);
-        if (it == certificates.end()) {
-          // Cannot find the separation certificate for this pair   of
-          // geometries.
-          auto diagram_context = diagram_->CreateDefaultContext();
-          auto plant_context =
-              &(plant_->GetMyMutableContextFromRoot(diagram_context.get()));
-          EXPECT_TRUE(FindInCollisionPosture(*plant_, geometry_pair,
-                                             plant_context, q_box_lower,
-                                             q_box_upper, std::nullopt)
-                          .has_value());
-        } else {
-          const auto& certificate_result = it->second;
-          CheckSeparationBySamples(
-              tester, *(this->diagram_), q_samples, certificate_result.a,
-              certificate_result.b, q_star,
-              SortedPair<geometry::GeometryId>(
-                  separating_plane.positive_side_geometry->id(),
-                  separating_plane.negative_side_geometry->id()));
-        }
-      }
-    }
   }
 }
 }  // namespace
